+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 #CHANGELOG- Sunderland Coding Assessment
 
+# Version 1.21.6.1 Release Notes
+  - Testing - Some minor changes to the splash screen instrumented test   
+
+
 # Version 1.21.6.0 Release Notes
-  - Improvement - Updated the splash screen transition delay to use the type-safe Duration class as opposed to Long delay, broke delay logic out into a more easily testable function and added test cases
+  - Testing - Broke splash screen delay logic out into a more easily testable function and added unit and instrumentation test cases
+  - Improvement - Updated the splash screen transition delay to use the type-safe Duration class as opposed to Long delay
   - Bug Fix - Splash screen gets stuck if the screen is turned off while the coroutine is running
   - General - Updated dependencies